small visual changes
+removed colours from buttons
+option button is now located relative to bottom of the window
+ price is decimal
+ green highlight when project has ready status
+ removed shippment column

+++ Nesdesign/Models/Offer.cs
@@ -78,7 +78,7 @@ namespace Nesdesign.Models
 
         [ObservableProperty]
         [Column("price")]
-        private int? price =100;
+        private decimal? price =100;
 
         [ObservableProperty]
         [Column("shipment")]
